print size
